chore: relocate utils' related stuff
