+++ apps/mobile/src/stacks/Dashboard/index.tsx
@@ -1,7 +1,7 @@
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
 import HomeStack from 'stacks/Home';
 import SettingStack from 'stacks/Setting';
-import { noHeaderNavigation } from 'utils/helper';
+import { noHeaderNavigation } from 'utils/constants';
 import { useNotificationPermissionRequest } from 'utils/hooks';
 import type { DashboardParamList } from 'utils/navigation';
