+++ apps/mobile/src/utils/constants.ts
@@ -1,1 +1,5 @@
+export const noHeaderNavigation = {
+	headerShown: false,
+};
+
 export const tabBarHeight = 52;
