+++ apps/mobile/src/utils/format.ts
@@ -14,3 +14,13 @@ export const formatQuote = (
 	if (!value) return fallbackDisplay;
 	return `${numeral(value).format(format)}${suffix}`;
 };
+
+export const removeRedundantCharacters = (text: string) =>
+	text.replaceAll('  ', '').replaceAll('\u0000', '').trim();
+
+export const convertDateToReadable = (date: Date) =>
+	date.toLocaleDateString('en-US', {
+		year: 'numeric',
+		month: 'long',
+		day: 'numeric',
+	});
